+++ ShopManagement/Program.cs
@@ -2,10 +2,10 @@ using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
-using PersonalWebsite.DatabaseModel;
 using System;
+using ShopManagement.DatabaseModel;
 
-namespace PersonalWebsite
+namespace ShopManagement
 {
     public class Program
     {
